Fix : renaming response class

+++ service-data/src/main/java/com/virnect/data/dto/response/guest/GuestInfoResponse.java
@@ -16,7 +16,7 @@ public class GuestInfoResponse {
 	private String workspaceId;
 	private String accessToken;
 	private String refreshToken;
-	private int expireIn;
+	private long expireIn;
 	private RoomDetailInfoResponse roomInfoResponse;
 
 	@Override
